Refactors scores to use Substance entity
Updates OnsiteScore and GroundwaterScore entities to use the Substance entity instead of Chemical directly.

This change simplifies the data model and ensures that all relevant substance information is readily available when calculating scores. It also adds validation to prevent duplicate substance assignments for scoring within a facility.

Removes ChemicalsSelectList Async and GetChemicalsFromSubstanceAsync methods.

+++ FMS/Pages/Substance/Delete.cshtml.cs
@@ -14,10 +14,14 @@ namespace FMS.Pages.Substance
     public class DeleteModel : PageModel
     {
         private readonly ISubstanceRepository _repository;
+        private readonly IGroundwaterScoreRepository _groundwaterScoreRepository;
+        private readonly IOnsiteScoreRepository _onsiteScoreRepository;
 
-        public DeleteModel(ISubstanceRepository repository)
+        public DeleteModel(ISubstanceRepository repository, IGroundwaterScoreRepository groundwaterScoreRepository, IOnsiteScoreRepository onsiteScoreRepository)
         {
             _repository = repository;
+            _groundwaterScoreRepository = groundwaterScoreRepository;
+            _onsiteScoreRepository = onsiteScoreRepository;
         }
 
         [BindProperty]
@@ -52,7 +56,16 @@ namespace FMS.Pages.Substance
 
         public async Task<IActionResult> OnPostAsync()
         {
+            if (await _groundwaterScoreRepository.SubstanceExistsInGroundwaterScoreAsync(Id, FacilityId) ||
+                await _onsiteScoreRepository.SubstanceExistsInOnsiteScoreAsync(Id, FacilityId))
+            {
+                TempData?.SetDisplayMessage(Context.Danger, "Cannot delete Substance because it is referenced by existing scores.");
+                ActiveTab = "Substances";
+                return RedirectToPage("/Facilities/Details", new { id = FacilityId, hr = Guid.Empty });
+            }
+
             await _repository.DeleteSubstanceAsync(Id);
+
             TempData?.SetDisplayMessage(Context.Success, "Substance deleted.");
             ActiveTab = "Substances";
             return RedirectToPage("/Facilities/Details", new { id = FacilityId, hr = Guid.Empty });
